--- a/benchmark_test.go
+++ b/benchmark_test.go
@@ -1,8 +1,10 @@
 package ltsvlog_test
 
 import (
+	"bytes"
 	"errors"
 	"io/ioutil"
+	"math"
 	"os"
 	"testing"
 	"time"
@@ -74,5 +76,21 @@ func BenchmarkErrWithUTCTime(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		err = run()
 		logger.Err(err)
+	}
+}
+
+func BenchmarkEvent_Float32(b *testing.B) {
+	buf := new(bytes.Buffer)
+	logger := ltsvlog.NewLTSVLogger(buf, true, ltsvlog.SetTimeLabel(""))
+	for i := 0; i < b.N; i++ {
+		logger.Info().Float32("max_float32", math.MaxFloat32).Float32("smallest_nonzero_float32", math.SmallestNonzeroFloat32).Log()
+	}
+}
+
+func BenchmarkEvent_Float64(b *testing.B) {
+	buf := new(bytes.Buffer)
+	logger := ltsvlog.NewLTSVLogger(buf, true, ltsvlog.SetTimeLabel(""))
+	for i := 0; i < b.N; i++ {
+		logger.Info().Float64("max_float64", math.MaxFloat64).Float64("smallest_nonzero_float64", math.SmallestNonzeroFloat64).Log()
 	}
 }