--- a/error.go
+++ b/error.go
@@ -182,7 +182,7 @@ func (e *Error) Float32(label string, value float32) *Error {
 	e.buf = append(e.buf, '\t')
 	e.buf = append(e.buf, label...)
 	e.buf = append(e.buf, ':')
-	e.buf = append(e.buf, strconv.FormatFloat(float64(value), 'g', -1, 32)...)
+	e.buf = strconv.AppendFloat(e.buf, float64(value), 'g', -1, 32)
 	return e
 }
 
@@ -191,7 +191,7 @@ func (e *Error) Float64(label string, value float64) *Error {
 	e.buf = append(e.buf, '\t')
 	e.buf = append(e.buf, label...)
 	e.buf = append(e.buf, ':')
-	e.buf = append(e.buf, strconv.FormatFloat(value, 'g', -1, 64)...)
+	e.buf = strconv.AppendFloat(e.buf, value, 'g', -1, 64)
 	return e
 }