```
go test -count=10 -run=NONE -bench=BenchmarkEvent_Float -benchmem -cpuprofile=cpu-old.prof | tee old.log
goos: linux
goarch: amd64
pkg: github.com/hnakamur/ltsvlog
BenchmarkEvent_Float32-2   	 2000000	       692 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       708 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       695 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       723 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       698 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       706 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       713 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       691 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       697 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float32-2   	 2000000	       697 ns/op	     227 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       685 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       676 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       668 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       677 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       671 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       676 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       655 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       678 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       678 ns/op	     269 B/op	       4 allocs/op
BenchmarkEvent_Float64-2   	 2000000	       669 ns/op	     269 B/op	       4 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	41.864s

go test -count=10 -run=NONE -bench=BenchmarkEvent_Float -benchmem -cpuprofile=cpu-new.prof | tee new.log
goos: linux
goarch: amd64
pkg: github.com/hnakamur/ltsvlog
BenchmarkEvent_Float32-2   	 3000000	       547 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       538 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       541 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       532 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       541 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       536 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       539 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       533 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       534 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float32-2   	 3000000	       537 ns/op	     190 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       496 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       510 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       506 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       508 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       506 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       510 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       508 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       509 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       506 ns/op	     220 B/op	       0 allocs/op
BenchmarkEvent_Float64-2   	 3000000	       504 ns/op	     220 B/op	       0 allocs/op
PASS
ok  	github.com/hnakamur/ltsvlog	42.029s

benchstat old.log new.log | tee stat.log
name             old time/op    new time/op    delta
Event_Float32-2     702ns ± 3%     538ns ± 2%   -23.39%  (p=0.000 n=10+10)
Event_Float64-2     675ns ± 1%     507ns ± 1%   -24.86%  (p=0.000 n=9+9)

name             old alloc/op   new alloc/op   delta
Event_Float32-2      227B ± 0%      190B ± 0%   -16.30%  (p=0.000 n=10+10)
Event_Float64-2      269B ± 0%      220B ± 0%   -18.22%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
Event_Float32-2      4.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
Event_Float64-2      4.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```